# backend/api/server.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// GetLaptops Liste aller Laptops abrufen
	// (GET /laptops)
	GetLaptops(ctx echo.Context, params GetLaptopsParams) error
	// PostLaptops Neuen Laptop zur Reparatur erfassen
	// (POST /laptops)
	PostLaptops(ctx echo.Context) error
	// DeleteLaptopsId Reparierten oder ausgemusterten Laptop löschen
	// (DELETE /laptops/{id})
	DeleteLaptopsId(ctx echo.Context, id int) error
	// GetLaptopsId Einzelnen Laptop samt Logs abrufen
	// (GET /laptops/{id})
	GetLaptopsId(ctx echo.Context, id int) error
	// PutLaptopsId Laptop-Daten oder Fehlerstatus aktualisieren
	// (PUT /laptops/{id})
	PutLaptopsId(ctx echo.Context, id int) error
	// PostLaptopsIdLogs IT-Log-Eintrag hinzufügen
	// (POST /laptops/{id}/logs)
	PostLaptopsIdLogs(ctx echo.Context, id int) error
	// DeleteLaptopsIdLogsLogId IT-Log-Eintrag löschen
	// (DELETE /laptops/{id}/logs/{logId})
	DeleteLaptopsIdLogsLogId(ctx echo.Context, id int, logId int) error
	// PutLaptopsIdLogsLogId Existierenden IT-Log-Eintrag aktualisieren
	// (PUT /laptops/{id}/logs/{logId})
	PutLaptopsIdLogsLogId(ctx echo.Context, id int, logId int) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// GetLaptops converts echo context to params.
func (w *ServerInterfaceWrapper) GetLaptops(ctx echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetLaptopsParams
	// ------------- Optional query parameter "fehler" -------------

	err = runtime.BindQueryParameterWithOptions("form", true, false, "fehler", ctx.QueryParams(), &params.Fehler, runtime.BindQueryParameterOptions{Type: "string", Format: ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter fehler: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetLaptops(ctx, params)
	return err
}

// PostLaptops converts echo context to params.
func (w *ServerInterfaceWrapper) PostLaptops(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostLaptops(ctx)
	return err
}

// DeleteLaptopsId converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteLaptopsId(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id int

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteLaptopsId(ctx, id)
	return err
}

// GetLaptopsId converts echo context to params.
func (w *ServerInterfaceWrapper) GetLaptopsId(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id int

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.GetLaptopsId(ctx, id)
	return err
}

// PutLaptopsId converts echo context to params.
func (w *ServerInterfaceWrapper) PutLaptopsId(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id int

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PutLaptopsId(ctx, id)
	return err
}

// PostLaptopsIdLogs converts echo context to params.
func (w *ServerInterfaceWrapper) PostLaptopsIdLogs(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id int

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostLaptopsIdLogs(ctx, id)
	return err
}

// DeleteLaptopsIdLogsLogId converts echo context to params.
func (w *ServerInterfaceWrapper) DeleteLaptopsIdLogsLogId(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id int

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// ------------- Path parameter "logId" -------------
	var logId int

	err = runtime.BindStyledParameterWithOptions("simple", "logId", ctx.Param("logId"), &logId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter logId: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.DeleteLaptopsIdLogsLogId(ctx, id, logId)
	return err
}

// PutLaptopsIdLogsLogId converts echo context to params.
func (w *ServerInterfaceWrapper) PutLaptopsIdLogsLogId(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "id" -------------
	var id int

	err = runtime.BindStyledParameterWithOptions("simple", "id", ctx.Param("id"), &id, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter id: %s", err))
	}

	// ------------- Path parameter "logId" -------------
	var logId int

	err = runtime.BindStyledParameterWithOptions("simple", "logId", ctx.Param("logId"), &logId, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "integer", Format: "", ValueIsUnescaped: ctx.Request().URL.RawPath == ""})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter logId: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PutLaptopsIdLogsLogId(ctx, id, logId)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlersOptions configures RegisterHandlersWithOptions.
type RegisterHandlersOptions struct {
	// BaseURL is prepended to every registered path so the API can be served
	// under a prefix.
	BaseURL string
	// OperationMiddlewares lets the caller attach per-operation middleware at
	// registration time. The map key is the OpenAPI `operationId` value as it
	// appears in the spec (the raw, un-normalized form). Operations that have
	// no entry are registered with no extra middleware. A nil map disables
	// per-operation middleware entirely.
	OperationMiddlewares map[string][]echo.MiddlewareFunc
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{})
}

// RegisterHandlersWithBaseURL registers handlers and prepends BaseURL to the
// paths so the API can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{BaseURL: baseURL})
}

// RegisterHandlersWithOptions registers handlers using the supplied options,
// including any per-operation middleware.
func RegisterHandlersWithOptions(router EchoRouter, si ServerInterface, options RegisterHandlersOptions) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.GET(options.BaseURL+"/laptops", wrapper.GetLaptops, options.OperationMiddlewares["GetLaptops"]...)
	router.POST(options.BaseURL+"/laptops", wrapper.PostLaptops, options.OperationMiddlewares["PostLaptops"]...)
	router.DELETE(options.BaseURL+"/laptops/:id", wrapper.DeleteLaptopsId, options.OperationMiddlewares["DeleteLaptopsId"]...)
	router.GET(options.BaseURL+"/laptops/:id", wrapper.GetLaptopsId, options.OperationMiddlewares["GetLaptopsId"]...)
	router.PUT(options.BaseURL+"/laptops/:id", wrapper.PutLaptopsId, options.OperationMiddlewares["PutLaptopsId"]...)
	router.POST(options.BaseURL+"/laptops/:id/logs", wrapper.PostLaptopsIdLogs, options.OperationMiddlewares["PostLaptopsIdLogs"]...)
	router.DELETE(options.BaseURL+"/laptops/:id/logs/:logId", wrapper.DeleteLaptopsIdLogsLogId, options.OperationMiddlewares["DeleteLaptopsIdLogsLogId"]...)
	router.PUT(options.BaseURL+"/laptops/:id/logs/:logId", wrapper.PutLaptopsIdLogsLogId, options.OperationMiddlewares["PutLaptopsIdLogsLogId"]...)

}